group by conflict
